- Set training data.
- Added collision responses to the neural network.
- Trained neural network and stored trained version.

Related work items: #41

+++ Slithering/stdafx.h
@@ -121,7 +121,7 @@
 // Global namespace
 namespace app
 {
-	typedef std::uint32_t Entity;
+	typedef std::uint64_t Entity;
 	typedef entt::Registry<Entity> Registry;
 	namespace inp
 	{
